plugin/runtime/requirements.txt
# Optional: use .yaml lexicons directly. Without it, the bundled .json lexicons are used (no install needed).
pyyaml>=5.1
